fix: added cancellation logic to the `Promise.race` in `Some` function (#110)
## Explanation

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->
In the `Some` function (auth-network-utils package), we used
`Promise.race` to enforce the network request timeout behaviour.

However, we do not add the cancellation logic to the `timeoutPromise`
when another promise is settled before the timeout promise duration.

This results in `dangling promise`. It currently affects in `test cases`
(the tests didn't finish util the `timeout` promise is rejected, which
is 10 seconds) and can potentially cause confusion and bugs.

## References

<!--
Are there any issues that this pull request is tied to?
Are there other links that reviewers should consult to understand these
changes better?
Are there client or consumer pull requests to adopt any breaking
changes?

For example:

* Fixes #12345
* Related to #67890
-->
* Fixes https://github.com/MetaMask/toprf-secure-backup/issues/109

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've highlighted breaking changes using the "BREAKING" category
above as appropriate
- [ ] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes