Extract and process `Rate Limit Errors` from the ToprfEvalResponse results (#120)
## Explanation

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->

Currently, SSS servers rate limit the `TOPRFEvalRequest` upon the third
attempts, however the servers do not respond with the error but include
the `RateLimitDetails` in the success result.
It's the SDK's job to correctly identify the Toprf Eval success/failure
and throw relevant rate-limit error.
In the existing implementation, we do not read the ratelimit data in the
results, and we rely on the error response from the SSS servers, hence
resulting the in-adequate rate limit results.

This PR addresses this issue by
- Check the `TOPRFEvalResponse` for `RateLimitError` extraction whenever
the derivation of encryption key fails.
- If the `ratelimit` data presents in the server response, throw the
`RateLimit` Error

## References

<!--
Are there any issues that this pull request is tied to?
Are there other links that reviewers should consult to understand these
changes better?
Are there client or consumer pull requests to adopt any breaking
changes?

For example:

* Fixes #12345
* Related to #67890
-->

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've highlighted breaking changes using the "BREAKING" category
above as appropriate
- [ ] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes